Show date and hour on customer conversation messages

+++ customers/src/main/java/app/botia/android/customers/api/dto/CustomerMessageDTO.kt
@@ -7,7 +7,7 @@ data class CustomerMessageDTO(
     @field:Json(name = "received_message") val customerAnswer: String,
     @field:Json(name = "bot_answer") val userAnswer: String,
     @field:Json(name = "message_id") val messageId: String,
-    val timestamp: Long,
+    val timestamp: Int,
     @field:Json(name = "customer_phone") val customerPhone: String,
     @field:Json(name = "messaging_product") val messagingProduct: String, // whatsapp, facebook or instagram
     @field:Json(name = "function_name") val functionName: String?, // The name of the ChatGPT function triggered to answer this
